Add ruby ctags script
Based off tpope's wonderful "effortless ctags" post, this introduces a
script that is meant to be called via git hooks. The intended usage of
this script is that it's called in the background, and only one instance
of it runs at a time. This is because the main repository I work on is a
really big Ruby repository.