Use git ls-files to drive create-symlinks
This changes the linking strategy to use git ls-files to figure out what
to symlink. It's better because it doesn't make any assumptions about
the bin directory, and the .config stuff will just work. Any future
directory that needs to be symlinked should be handled by this.

To do this, I needed to add a list of exceptions (i.e. things I won't be
copying over). The list of exceptions should always be pretty small, so
this should work for a while.